+++ app/src/main/java/com/teamwss/websoso/ui/main/myPage/myActivity/MyActivityFragment.kt
@@ -2,29 +2,28 @@ package com.teamwss.websoso.ui.main.myPage.myActivity
 
 import android.os.Bundle
 import android.view.View
+import android.widget.TextView
 import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
 import com.teamwss.websoso.R
 import com.teamwss.websoso.common.ui.base.BaseFragment
 import com.teamwss.websoso.databinding.FragmentMyActivityBinding
 import com.teamwss.websoso.ui.activityDetail.ActivityDetailActivity
+import com.teamwss.websoso.ui.feedDetail.FeedDetailActivity
 import com.teamwss.websoso.ui.main.myPage.MyPageViewModel
 import com.teamwss.websoso.ui.main.myPage.myActivity.adapter.MyActivityAdapter
 import com.teamwss.websoso.ui.main.myPage.myActivity.model.UserProfileModel
+import com.teamwss.websoso.ui.novelDetail.NovelDetailActivity
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class MyActivityFragment :
-    BaseFragment<FragmentMyActivityBinding>(R.layout.fragment_my_activity) {
+    BaseFragment<FragmentMyActivityBinding>(R.layout.fragment_my_activity),ActivityItemClickListener {
     private val myActivityViewModel: MyActivityViewModel by viewModels()
     private val myPageViewModel: MyPageViewModel by activityViewModels()
-    private val activityItemClickHandler: ActivityItemClickHandler by lazy {
-        ActivityItemClickHandler(requireContext(), myActivityViewModel)
-    }
     private val myActivityAdapter: MyActivityAdapter by lazy {
-        MyActivityAdapter(activityItemClickHandler)
+        MyActivityAdapter(this)
     }
-    private var userProfile: UserProfileModel? = null
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -48,14 +47,9 @@ class MyActivityFragment :
                     nickname = myProfileEntity.nickname,
                     avatarImage = myProfileEntity.avatarImage,
                 )
-                setupUserProfile(userProfile)
+                myActivityAdapter.setUserProfile(userProfile)
             }
         }
-    }
-
-    fun setupUserProfile(userProfile: UserProfileModel) {
-        this.userProfile = userProfile
-        myActivityAdapter.setUserProfile(userProfile)
     }
 
     private fun onMyActivityDetailButtonClick() {
@@ -65,6 +59,34 @@ class MyActivityFragment :
             }
             startActivity(intent)
         }
+    }
+
+    override fun onContentClick(feedId: Long) {
+        startActivity(FeedDetailActivity.getIntent(requireContext(), feedId))
+    }
+
+    override fun onNovelInfoClick(novelId: Long) {
+        startActivity(NovelDetailActivity.getIntent(requireContext(), novelId))
+    }
+
+    override fun onLikeButtonClick(view: View, feedId: Long) {
+        val likeCountTextView: TextView = view.findViewById(R.id.tv_my_activity_thumb_up_count)
+        val currentLikeCount = likeCountTextView.text.toString().toInt()
+
+        val updatedLikeCount: Int = if (view.isSelected) {
+            if (currentLikeCount > 0) currentLikeCount - 1 else 0
+        } else {
+            currentLikeCount + 1
+        }
+
+        likeCountTextView.text = updatedLikeCount.toString()
+        view.isSelected = !view.isSelected
+
+        myActivityViewModel.updateActivityLike(view.isSelected, feedId, updatedLikeCount)
+    }
+
+    override fun onMoreButtonClick(view: View, feedId: Long) {
+        // TODO 팝업메뉴 수정 and 차단
     }
 
     companion object {
